moar

+++ exemples/sph_weak_scale_test.py
@@ -71,7 +71,7 @@ setup.apply_setup(
     gen,
     insert_step=int(scheduler_split_val / 4),
     msg_count_limit=64,
-    msg_size_limit=scheduler_split_val /2,
+    msg_size_limit=int(scheduler_split_val /2),
     do_setup_log=True,
 )
